Feat: Added some extra calculators, implemented BigDecimalFormatter, refactored code

+++ app/src/main/java/com/cobaltumapps/simplecalculator/data/extra/calculator_unit/ExtraUnitInfo.kt
@@ -1,9 +1,11 @@
 package com.cobaltumapps.simplecalculator.data.extra.calculator_unit
 
+import java.math.BigDecimal
+
 /** Data-класс содержащий информацию о единице измерения для UI */
 
 data class ExtraUnitInfo(
     val unitName: String,
     val unitPreview: String?,
-    var unitValue: Float
+    var unitValue: BigDecimal
 )
